added screens and switching between them
